+++ core/evol/method/select/BestSelector.cpp
@@ -4,7 +4,7 @@
 
 namespace ea::method {
 
-void BestSelector::select(instance::Population& population, size_t size) {
+void BestSelector::select(instance::Population& population, uint32_t size) {
   if (population.size() > size) {
     std::sort(population.begin(), population.end(), [](auto& a, auto& b) { return *a < *b; });
     population.resize(size);
